+++ src/modules/api.ts
@@ -6,6 +6,7 @@ import { Integrations } from './integrations';
 import { Webhooks } from './webhooks';
 import { DataStores } from './datastores';
 import { Issues } from './issues';
+import { Datalab } from './datalab';
 
 export class API {
     auth: AuthType;
@@ -17,6 +18,7 @@ export class API {
     Webhooks;
     DataStores;
     Issues;
+    Datalab;
 
     constructor(auth: AuthType) {
         this.auth = auth;
@@ -27,6 +29,7 @@ export class API {
         this.Webhooks = Webhooks(this.internalApi);
         this.DataStores = DataStores(this.internalApi);
         this.Issues = Issues(this.internalApi);
+        this.Datalab = Datalab(this.internalApi);
     }
 
     private _setup = () => {